Small fixes (#62)
* Fix recapContainer reader width (#59)

* Refactor extractChapter content:

- Added explaining comments
- Made the loop less error prone
- Improved readability

* Fix unsafe DOM manipulation

* Update recap word count display

There is a difference in word amount calculation between extracting the content and creating the recap, so I reverted it for now.

* Fix preserving styles in the last paragraph of the loop

Previoulsy, the loop would just take the textContent of the last paragraph
and split it according to the reminaing count. This removed any styles an
author might apply, e.g. fonts, so I implemented a way to preserve those
styles.

* Updated startUrl for better testing